Refactor resource-request-controller tests and move controller to the liqo-controller-manager package
