+++ pkg/liqo-controller-manager/resource-request-controller/interfaces/updaterInterface.go
