+++ pkg/liqo-controller-manager/resource-request-controller/testutils/doc.go
@@ -1,0 +1,2 @@
+// Package testutils contains utility functions used by the unit tests of the resource-request-controller
+package testutils
